--- a/src/Autofac/Core/Resolving/Middleware/CircularDependencyDetectorMiddleware.cs
+++ b/src/Autofac/Core/Resolving/Middleware/CircularDependencyDetectorMiddleware.cs
@@ -68,7 +68,16 @@ namespace Autofac.Core.Resolving.Middleware
 
             if (activationDepth > _maxResolveDepth)
             {
+#if NETSTANDARD2_1
+                // In .NET Standard 2.1 we will try and keep going until we run out of stack space.
+                if (!RuntimeHelpers.TryEnsureSufficientExecutionStack())
+                {
+                    throw new DependencyResolutionException(string.Format(CultureInfo.CurrentCulture, CircularDependencyDetectorMessages.MaxDepthExceeded, context.Service));
+                }
+#else
+                // Pre .NET Standard 2.1 we just end at 50.
                 throw new DependencyResolutionException(string.Format(CultureInfo.CurrentCulture, CircularDependencyDetectorMessages.MaxDepthExceeded, context.Service));
+#endif
             }
 
             var requestStack = context.Operation.RequestStack;