… src/Autofac/Core/Resolving/SegmentedStackResources.Designer.cs
… test/Autofac.Test/Core/Resolving/SegmentedStackTests.cs